rename icon file name
