+++ webapp/channels/src/components/widgets/icons/entra_id_icon.tsx
